IMP - Refactor nav files
