+++ pages/_app.tsx
@@ -10,7 +10,7 @@ import '../public/bs-icons.css';
 import '../public/index.css';
 import {isProduction} from '../server/utils/misc';
 import {Footer} from '../src/components/elements/footer';
-import {Navigation} from '../src/components/elements/nav/main';
+import {Navigation} from '../src/components/nav/main';
 import {SiteAlert} from '../src/components/pages/siteAlert';
 import {GlobalAlert} from '../src/components/pages/stateAlert';
 import {AppReactContext} from '../src/context/app/main';
