+++ src/components/nav/elements/dropdownItem.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import NavDropdown from 'react-bootstrap/NavDropdown';
 
-import {useNextRouter} from '../../../../utils/router';
+import {useNextRouter} from '../../../utils/router';
 import {NavProps} from './types';
