+++ src/components/nav/elements/item.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import Nav from 'react-bootstrap/Nav';
 
-import {useNextRouter} from '../../../../utils/router';
+import {useNextRouter} from '../../../utils/router';
 import {NavProps} from './types';
